+++ sophon/src/arch/aarch64/exception.rs
@@ -1,7 +1,6 @@
 use crate::arch::{aarch64::context::*, *};
 use crate::modules::INTERRUPT;
 use crate::modules::PROCESS_MANAGER;
-use crate::task::proc::TaskExt;
 use core::arch::{asm, global_asm};
 use cortex_a::{asm::barrier, registers::*};
 use tock_registers::interfaces::{Readable, Writeable};
@@ -67,10 +66,7 @@ unsafe fn is_el0(frame: &ExceptionFrame) -> bool {
 pub unsafe extern "C" fn handle_exception(exception_frame: &mut ExceptionFrame) {
     // log!("Exception received");
     let privileged = !is_el0(exception_frame);
-    PROCESS_MANAGER
-        .current_task()
-        .unwrap()
-        .get_context::<AArch64Context>()
+    AArch64Context::of(&*PROCESS_MANAGER.current_task().unwrap())
         .push_exception_frame(exception_frame);
     let exception = get_exception_class();
     match exception {
@@ -104,10 +100,8 @@ pub unsafe extern "C" fn handle_exception(exception_frame: &mut ExceptionFrame) 
         _ => panic_for_unhandled_exception(exception_frame),
     }
     // Note: `Task::current()` must be dropped before calling `return_to_user`.
-    let context = PROCESS_MANAGER
-        .current_task()
-        .unwrap()
-        .get_context::<AArch64Context>() as *const AArch64Context;
+    let context =
+        AArch64Context::of(&*PROCESS_MANAGER.current_task().unwrap()) as *const AArch64Context;
     (*context).return_to_user();
 }
 
@@ -147,19 +141,14 @@ unsafe fn panic_for_unhandled_exception(exception_frame: *mut ExceptionFrame) ->
 pub unsafe extern "C" fn handle_interrupt(exception_frame: &mut ExceptionFrame) {
     INTERRUPT.interrupt_begin();
     let irq = INTERRUPT.get_active_irq().unwrap();
-    PROCESS_MANAGER
-        .current_task()
-        .unwrap()
-        .get_context::<AArch64Context>()
+    AArch64Context::of(&*PROCESS_MANAGER.current_task().unwrap())
         .push_exception_frame(exception_frame);
     super::super::handle_irq(irq);
     INTERRUPT.interrupt_end();
     ::core::sync::atomic::fence(::core::sync::atomic::Ordering::SeqCst);
     // Note: `Task::current()` must be dropped before calling `return_to_user`.
-    let context = PROCESS_MANAGER
-        .current_task()
-        .unwrap()
-        .get_context::<AArch64Context>() as *const AArch64Context;
+    let context =
+        AArch64Context::of(&*PROCESS_MANAGER.current_task().unwrap()) as *const AArch64Context;
     (*context).return_to_user();
 }
