+++ sophon/src/utils/testing.rs
@@ -6,7 +6,6 @@ pub use testing::{Test, Tests};
 use crate::{
     arch::{Arch, TargetArch},
     modules::PROCESS_MANAGER,
-    task::MMState,
 };
 
 static BOOT_TESTS: RwLock<Tests> = RwLock::new(Tests::new("boot"));
@@ -37,7 +36,7 @@ pub fn run_boot_tests() {
 
 pub fn start_kernel_test_runner() {
     assert!(cfg!(sophon_test));
-    let _proc = PROCESS_MANAGER.spawn(box KernelTestRunner, MMState::new());
+    let _proc = PROCESS_MANAGER.spawn(box KernelTestRunner);
 }
 
 pub struct KernelTestRunner;
